feat: investor project screen
