--- src/pages/loginAdm/styles.ts
@@ -1,5 +1,0 @@
-export const selectStyle = 'flex bg-optionWhite h-12 w-471 rounded-full direction-row items-center justify-between px-[.5rem] mt-5';
-export const divGeneral = 'flex justify-center items-center bg-white h-screen w-full grid direction-column';
-export const textTitle = 'font-semibold text-4xl mt-3';
-export const buttonForgotPassword = 'flex justify-end w-full text-greenText65 font-bold text-xs underline mt-4';
-export const buttonRegister = 'flex items-center justify-center w-full text-greenText font-bold text-xs underline mt-4';
